Move shred_version module to sdk/
