+++ sdk/src/shred_version.rs
@@ -1,6 +1,23 @@
-use crate::shred::Shred;
-use solana_runtime::hard_forks::HardForks;
-use solana_sdk::hash::{extend_and_hash, Hash};
+use solana_sdk::{
+    hard_forks::HardForks,
+    hash::{extend_and_hash, Hash},
+};
+
+pub fn version_from_hash(hash: &Hash) -> u16 {
+    let hash = hash.as_ref();
+    let mut accum = [0u8; 2];
+    hash.chunks(2).for_each(|seed| {
+        accum
+            .iter_mut()
+            .zip(seed)
+            .for_each(|(accum, seed)| *accum ^= *seed)
+    });
+    // convert accum into a u16
+    let version = ((accum[0] as u16) << 8) | accum[1] as u16;
+
+    // ensure version is never zero, to avoid looking like an uninitialized version
+    version.saturating_add(1)
+}
 
 pub fn compute_shred_version(genesis_hash: &Hash, hard_forks: Option<&HardForks>) -> u16 {
     use byteorder::{ByteOrder, LittleEndian};
@@ -15,7 +32,7 @@ pub fn compute_shred_version(genesis_hash: &Hash, hard_forks: Option<&HardForks>
         }
     }
 
-    Shred::version_from_hash(&hash)
+    version_from_hash(&hash)
 }
 
 #[cfg(test)]
